Removing console.log Moving Files on InstallPluginDialog
